requirements.txt
mkdocs
pip
python-dotenv
ruff
-e .
pandas
numpy
matplotlib
torch
torchvision
pillow
tqdm
scikit-learn